+++ eo_man/view/serial_communication_bar.py
@@ -221,9 +221,9 @@ class SerialConnectionBar():
         self.cb_device_type.config(state="readonly")
         try:
             self.cb_serial_ports['values'] = self.endpoint_list[get_gateway_type_by_name(self.cb_device_type.get())]
+            self.b_connect.config(state=NORMAL)
             if len(self.cb_serial_ports['values']) > 0:
                 self.cb_serial_ports.set(self.cb_serial_ports['values'][0])
-                self.b_connect.config(state=NORMAL)
                 self.cb_serial_ports.config(state=NORMAL)
             else:
                 # self.b_connect.config(state=DISABLED)